Swagger自動生成コード追加
