+++ src/main/kotlin/com/myblog/dev/controller/AuthController.kt
@@ -1,14 +1,16 @@
 package com.myblog.dev.controller
 
 import com.myblog.dev.auth.JwtUtil
-import com.myblog.dev.service.UserRegisterService
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
+import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.ResponseEntity
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
-import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -16,14 +18,20 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/auth")
+@Tag(name = "AUTH", description = "認証管理API")
 class AuthController(
     private val authenticationManager: AuthenticationManager,
     private val userDetailsService: UserDetailsService,
     private val jwtUtil: JwtUtil,
-    private val passwordEncoder: PasswordEncoder,
-    private val userRegisterService: UserRegisterService,
 ) {
 
+    @Operation(summary = "ログイン", description = "ログインする")
+    @ApiResponses(
+        value = [
+            ApiResponse(responseCode = "200", description = "OK"),
+            ApiResponse(responseCode = "400", description = "Bad Request"),
+        ]
+    )
     @PostMapping("/login")
     fun login(@RequestBody authRequest: LoginRequest): ResponseEntity<LoginResponse> {
         val authentication = authenticationManager.authenticate(
@@ -38,21 +46,6 @@ class AuthController(
         SecurityContextHolder.getContext().authentication = authentication
 
         return ResponseEntity.ok(LoginResponse(jwt))
-    }
-
-    @PostMapping("/register")
-    fun register(@RequestBody request: RegisterUserRequest): ResponseEntity<RegisterUserResponse> {
-        // ユーザー登録処理を実装。ユーザー情報を保存した後、JWTを返す。
-
-        val user = User(
-            username = request.email,
-            password = passwordEncoder.encode(request.password),
-            roles = listOf()
-        )
-        val jwt = jwtUtil.generateToken(userDetailsService.loadUserByUsername(user.username))
-
-        userRegisterService.register(form = request)
-        return ResponseEntity.ok(RegisterUserResponse(jwt))
     }
 }
